Move make_html to scripts, remove topcoffea/plotter
