+++ analysis/topEFT/mc_validation_plotter.py
@@ -16,7 +16,7 @@ from coffea import hist
 from topcoffea.modules.HistEFT import HistEFT
 from topcoffea.modules.paths import topcoffea_path
 from topcoffea.modules.YieldTools import YieldTools
-from topcoffea.plotter.make_html import make_html
+from topcoffea.scripts.make_html import make_html
 
 import topcoffea.modules.GetValuesFromJsons as getj
